refactor(reports): move markdown generation to separate ts file
